% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pnetm}
\alias{pnetm}
\title{Small example of a phonological network as an adjacency matrix}
\format{
Adjacency matrix representing an unweighted undirected graph with
  34 vertices and 96 edges. There are no self-loops.
}
\source{
Ying, Chan & Vitevitch, Michael. (2009). The Influence of the
  Phonological Neighborhood Clustering Coefficient on Spoken Word
  Recognition. Journal of experimental psychology. Human perception and
  performance. 35. 1934-49. 10.1037/a0016902.
}
\usage{
pnetm
}
\description{
Small example of a phonological network as an adjacency matrix
}
\keyword{datasets}
