% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_old.R
\name{spread}
\alias{spread}
\title{Spread of activation process to determine activation of network nodes.}
\usage{
spread(network, start_run, retention = 0.5, time = 10, decay = 0,
  suppress = 0)
}
\arguments{
\item{network}{Network where the spreading occurs. Must be specified. Must be an igraph object with a "name" property or an adjacency matrix.}

\item{start_run}{A non-empty dataframe with 'node' and 'activation' columns. Must be specified.}

\item{retention}{Proportion of activation that remains in the node from which activation is being spread from. Ranges from 0 to 1. Default is 0.5.}

\item{time}{Number of time steps to run the spread() function for. Default is 10.}

\item{decay}{Proportion of activation that is lost at each time step. Ranges from 0 to 1. Default is 0.}

\item{suppress}{Suppress nodes with a final activation of < x units at each time step to an activation value of 0. Default is 0.}
}
\value{
A compiled dataframe with 'node', 'activation' and 'time' columns showing the spread of activation in the network over time.
}
\description{
This function takes in a dataframe with 'node' and 'activation' columns, and
simulates the spread of activation among nodes in a specified network structure.
}
\examples{
#See Vignette for examples.
}
